Clean up and rename files
