+++ modules/geoscript/__init__.py
@@ -1,73 +1,9 @@
 #!/usr/bin/python3
 
-import bpy
+from .geofunction import geometry_function
+from .nodetrees import GeometryNodeTree
 
-from . import test_node_trees
-
-bl_info = {
-    "name": "GeoScript",
-    "author": "Lathreas",
-    "version": (0, 0, 1),
-    "blender": (3, 2, 2),
-    "location": "Unknown",
-    "description": "Create geometry modifiers using Python code",
-    "warning": "Experimental",
-    "category": "3D View",
-    "doc_url": "https://",
-    "tracker_url": "https://",
-}
-
-
-# Define an operator (button) that allows the user to run the script:
-class GeoscriptTestingOperator(bpy.types.Operator):
-    """Run GeoScript test functions"""
-
-    bl_idname = "geoscript.run_tests"
-    bl_label = "Run GeoScript test functions"
-
-    def execute(self, context):
-        test_geometry_modifier = test_node_trees.ExampleFunction(
-            "test_geometry_modifier"
-        )
-        test_normal_distribution = test_node_trees.NormalDistribution(
-            "common.normal_distribution"
-        )
-        test_tubercule = test_node_trees.Tubercule("tubercule")
-        return {"FINISHED"}
-
-
-class GeoscriptTestingPanel(bpy.types.Panel):
-    """Creates a Panel in the text editor window"""
-
-    bl_label = "Hello World Panel"
-    bl_space_type = "TEXT_EDITOR"
-    bl_category = "GeoScript"
-    bl_region_type = "UI"
-
-    def draw(self, context):
-        layout = self.layout
-
-        obj = context.object
-
-        row = layout.row()
-        row.label(text="Hello world!", icon="WORLD_DATA")
-
-        row = layout.row()
-        row.operator("geoscript.run_tests")
-
-
-def register():
-    bpy.utils.register_class(GeoscriptTestingOperator)
-    bpy.utils.register_class(GeoscriptTestingPanel)
-    print("Registered GeoScript Addon")
-    print("is debug: ", __debug__)
-
-
-def unregister():
-    bpy.utils.unregister_class(GeoscriptTestingOperator)
-    bpy.utils.unregister_class(GeoscriptTestingPanel)
-    print("Unregistered GeoScript Addon")
-
-
-if __name__ == "__main__":
-    register()
+__all__ = [
+    "geometry_function",
+    "GeometryNodeTree",
+]
