+++ modules/geoscript/test_node_trees.py
@@ -2,10 +2,9 @@
 
 import bpy
 
-from .geoscript import GeometryNodeTree
+from .nodetrees import GeometryNodeTree
 from .geofunction import geometry_function
 from .types import Vector3, Scalar, Geometry, Boolean, Object
-from .raytracing import raycast_with_attribute
 import math as m
 from . import math as g
 
@@ -128,20 +127,6 @@ class ExampleFunction(GeometryNodeTree):
         geo15 = geo14.get_convex_hull()
         vec1, vec2 = geo15.get_bounding_box_points()
 
-        rayhit = raycast_with_attribute(vec1, vec2, variable, geo15, variable, "FLOAT")
-        rayhit = raycast_with_attribute(vec1, vec2, variable, geo15, variable, "INT")
-        rayhit = raycast_with_attribute(vec1, vec2, variable, geo15, boolean, "BOOLEAN")
-        rayhit = raycast_with_attribute(
-            vec1, vec2, variable, geo15, vector1, "FLOAT_VECTOR"
-        )
-        # rayhit = raycast(vec1, vec2, variable, geo15, None, 'FLOAT_COLOR')
-
-        is_hit = rayhit.is_hit()
-        hit_position = rayhit.hit_position()
-        hit_normal = rayhit.hit_normal()
-        hit_distance = rayhit.hit_distance()
-        attribute = rayhit.attribute()
-
         variable2 = variable + 3.0
         variable3 = variable2 + variable
         variable4 = 4.0 + variable2
@@ -168,7 +153,6 @@ class ExampleFunction(GeometryNodeTree):
         abc2 = lerp(vector1, vector2, 1.0)
 
         self.OutputGeometry(geometry2, "Output Geometry")
-        self.OutputFloat(hit_distance, "Output Float")
         self.OutputFloat(
             variable8, "Float Output Name", attribute_domain="POINT", default_value=0.5
         )
